requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
websockets
jinja2
google-generativeai
itsdangerous
